Enhance data processing and upload functionality
- Added sequence counter reset option in data route for device restarts.
- Improved error messages for data processing failures.
- Updated UploadManager to log HTTP response bodies for better diagnostics.
- Fixed potential out-of-bounds access in update_config.cpp.
- Updated UploadManager's endpoint URLs for correct server address.

+++ src/update_config.cpp
@@ -24,7 +24,7 @@ namespace UpdateConfig {
 
                 int idx = 0;
                 int lastPos = 0;
-                while (idx < NUM_REGISTERS && lastPos < arrayStr.length()) {
+                while (idx < NUM_REGISTERS && (unsigned int)lastPos < arrayStr.length()) {
                     int comma = arrayStr.indexOf(',', lastPos);
                     String token = (comma == -1)
                         ? arrayStr.substring(lastPos)
